+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Changelog
 * Added Notification for Alert
 * Added alertRuleTags field to the graph panel
 * Added support for thresholds to graph panel
+* Added support for using gridPos for dashboard panels
 
 Changes
 -------